Circle 생성자 수정

+++ Hirearchy/Circle.cpp
@@ -6,7 +6,7 @@ Circle::Circle()
 {
 }
 
-Circle::Circle(const Point& c, int r)
+Circle::Circle(const Point& c, double r)
 	: center(c), rad(r)
 {
 }
